refactor: extract widget menu
